More refactoring of resnet_train
